+++ source/renderer/app/api/wallets/requests/transferFundsCalculateFee.js
@@ -1,0 +1,17 @@
+// @flow
+import type { RequestConfig } from '../../common/types';
+import type {
+  TransferFundsCalculateFeeRequest,
+  TransferFundsCalculateFeeResponse,
+} from '../types';
+import { request } from '../../utils/request';
+
+export const transferFundsCalculateFee = (
+  config: RequestConfig,
+  { sourceWalletId }: TransferFundsCalculateFeeRequest
+): Promise<TransferFundsCalculateFeeResponse> =>
+  request({
+    method: 'GET',
+    path: `/v2/byron-wallets/${sourceWalletId}/migrations`,
+    ...config,
+  });