+++ source/renderer/app/components/notifications/LegacyNotification.js
@@ -29,11 +29,18 @@ const messages = defineMessages({
     defaultMessage: '!!!Move all of the ada from this wallet',
     description: 'Move all ada action of legacy notification.',
   },
+  addWallet: {
+    id: 'wallet.legacy.notification.addWallet',
+    defaultMessage: '!!!Move all of the ada from this wallet',
+    description: 'Add wallet action of legacy notification.',
+  },
 });
 
 type Props = {
   onLearnMore: Function,
-  onMove: Function,
+  onTransferFunds: Function,
+  hasAnyWallets?: boolean,
+  onWalletAdd?: boolean,
 };
 
 @observer
@@ -44,11 +51,21 @@ export default class LegacyNotification extends Component<Props> {
 
   render() {
     const { intl } = this.context;
-    const { onLearnMore, onMove } = this.props;
+    const {
+      onLearnMore,
+      onTransferFunds,
+      hasAnyWallets,
+      onWalletAdd,
+    } = this.props;
     const title = intl.formatMessage(messages.title);
     const description = intl.formatMessage(messages.description);
     const actionLearnMore = intl.formatMessage(messages.actionLearnMore);
-    const actionMove = intl.formatMessage(messages.actionMove);
+
+    const buttonLabel = hasAnyWallets
+      ? intl.formatMessage(messages.actionMove)
+      : intl.formatMessage(messages.addWallet);
+
+    const buttonAction = hasAnyWallets ? onTransferFunds : onWalletAdd;
 
     return (
       <div className={styles.component}>
@@ -61,12 +78,14 @@ export default class LegacyNotification extends Component<Props> {
             onClick={onLearnMore}
             skin={ButtonSkin}
           />
-          <Button
-            className={styles.actionMove}
-            label={actionMove}
-            onClick={onMove}
-            skin={ButtonSkin}
-          />
+          {
+            <Button
+              className={styles.actionMove}
+              label={buttonLabel}
+              onClick={buttonAction}
+              skin={ButtonSkin}
+            />
+          }
         </div>
       </div>
     );